Refined the statistics engine - see description
The statistics engine is now capable of getting statistics of specific
properties for all files in a directory.
For example:
We want the following stats:
1. Number of ! in the comment
2. Number of :) in the comment
and so on.

The engine is now programmed to get whatever stat is required, provided
that there is a function that can handle the wanted property.